refactor: Extract user display and logout functionality into a new `UserMenu` client component.
